observation model bug fixed

+++ examples/cpp_models/ras/include/operator_model.h
@@ -16,7 +16,7 @@ public:
         acc_time_slope(0.25){
         }
 
-    OperatorModel(int _min_time, int _acc_time_min, double _acc_time_slope):
+    OperatorModel(int _min_time, double _acc_time_min, double _acc_time_slope):
         min_time(_min_time),
         acc_time_min(_acc_time_min),
         acc_time_slope(_acc_time_slope){
